Move 'clickOnText' to private folder
