+++ addon/-private/properties/click-on-text/acceptance.js
@@ -1,15 +1,13 @@
-import {
-  buildSelector,
-  simpleFindElementWithAssert
-} from '../../../helpers';
+import { simpleFindElementWithAssert } from '../../../helpers';
+import { buildSelector } from './helpers';
 
 /* global wait, click */
 
 export default function acceptanceClick(pageObjectNode, selector, options) {
-  wait().then(() => {
+  wait().then(function() {
     let fullSelector = buildSelector(pageObjectNode, selector, options);
 
-    // Run this to validate if the element exists and it is visible
+    // Run this to validate if the element exists
     simpleFindElementWithAssert(pageObjectNode, fullSelector, options);
 
     click(fullSelector, options.testContainer);
