+++ addon/index.js
@@ -1,7 +1,7 @@
 import { create }      from './create';                   export { create };
 import { collection }  from './collection';               export { collection };
-import { clickable }   from './-private/properties/clickable'; export { clickable };
-import { clickOnText } from './actions/click-on-text';    export { clickOnText };
+import { clickable }   from './-private/properties/clickable';        export { clickable };
+import { clickOnText } from './-private/properties/click-on-text';    export { clickOnText };
 import { fillable }    from './actions/fillable';         export { fillable }; export const selectable = fillable;
 import { visitable }   from './actions/visitable';        export { visitable };
 import { triggerable } from './actions/triggerable';      export { triggerable };
